Update Factorial.c
I made a few corrections to the code to make sure that it runs smoothly. At the beginning, the number that was entered was stored in "fac," but in line 11 the number 2 was put into fac. So, I made the number stored in "no." I also made fac=1 because you need to start at 1 when multiplying numbers, not at 2. On line 15, "w" was never used down below, so I replaced it with "no." There was also no semi-colon at the end of line 18, so I added that.